requirements.txt
fuzzywuzzy==0.18.0
gensim==4.2.0
pandas==1.5.1
Pillow==9.3.0
streamlit==1.14.0